Update mom v1.0 (2020/21) to mom v2.0 (2022)
I did a lot of work here; it would be difficult to document it all.
I first started this project in Fall 2020. That was mom v1.0.

This v2.0 is a revamped version. I worked on it in Summer 2022, and
I decided to put it on GitHub, because it is only now I am getting
comfortable with version control. Previously, I had been sending
emails to myself to backup the code.
From now on, I will only be using Git for version control.